requirements.txt
djangorestframework==3.12.4
djongo==1.3.4
